pydriver/requirements.txt
numpy
matplotlib
adafruit-circuitpython-mlx90640
python-socketio
eventlet

